smp_bruteforce/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
appdirs==1.4.4
attrs==25.3.0
bumble==0.0.214
cffi==2.0.0
click==8.2.1
cryptography==45.0.7
frozenlist==1.7.0
grpcio==1.74.0
humanize==4.13.0
idna==3.10
libusb-package==1.0.26.1
libusb1==3.3.1
multidict==6.6.4
platformdirs==4.4.0
prettytable==3.16.0
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.1
pycparser==2.23
pycryptodome==3.23.0
pyee==13.0.0
pyserial==3.5
pyserial-asyncio==0.6
pyusb==1.3.1
typing_extensions==4.15.0
wcwidth==0.2.13
websockets==13.1
yarl==1.20.1
